PR-fix : Moved the FPSLimiter back
Changed the FPS limiter and ApplicationQuit scripts to be inside the correct namespace.
