--- Assets/ProjectUtilities.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9ff12ef635d5b6648a0d75152db52cc2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
