+++ Assets/UGSSamples/FriendsSample/Scripts/Common/Utilities/FpsLimiter.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+namespace Unity.Services.Samples.Utilities
+{
+    // A simple class for limiting FPS for samples
+    public class FpsLimiter : MonoBehaviour
+    {
+        void Start()
+        {
+            Application.targetFrameRate = 60; // Used on non-PC platforms
+            QualitySettings.vSyncCount = 1; // Used on PC platforms
+        }
+    }
+}
